+++ packages/tools/src/index.ts
@@ -1,3 +1,0 @@
-import { capitalize } from './utils/capitalize';
-
-export { capitalize };