+++ sites/mytril.dev/mytril.config.js
@@ -1,6 +1,0 @@
-export default {
-	theme: {
-		defaultTheme: 'light',
-		colors: {}
-	}
-};