+++ sites/mytril.dev/src/routes/+layout.ts
@@ -1,8 +1,0 @@
-import { fetchPackageVersion } from '$lib/store/packages';
-
-export const load = async () => {
-	const packages = ['mytril', 'mytril-tools'];
-	await Promise.all(packages.map(fetchPackageVersion));
-
-	return {};
-};